requirements.txt
requests
tavily-python
feedparser
streamlit
python-dotenv